Modified SAML identity provider configuration to accept an addShadowUserOnLogin property, which is true by default, and determines whether to add new shadow users automatically or not.

+++ uaa/src/test/java/org/cloudfoundry/identity/uaa/integration/feature/SamlLoginIT.java
@@ -344,6 +344,7 @@ public class SamlLoginIT {
         if (null == userId) {
             return;
         }
+        if (userId == null) return;
 
         IntegrationTestUtils.deleteUser(zoneAdminToken, baseUrl, userId);
     }
